change goodsViewcontroller to goodsCollectionView
